+++ src/pages/Feedback.js
@@ -1,9 +1,12 @@
 import React from 'react';
+import Header from './Header';
 
 class Feedback extends React.Component {
   render() {
     return (
-      <div data-testid="feedback-text">Feedback</div>
+      <div data-testid="feedback-text">
+        <Header />
+      </div>
     );
   }
 }
